Pulled row and col validations out of .valid_move? method into their own separate methods:   .valid_row_dist? and .valid_col_dist
Added method to determine player direction in positive or negative values.

Added ability for .move_direction_calculation method to determin positive and negative distance as well as validate that the distance can only be 1 and 2 squares.

Added tests for all of the methods I added and made the .valid_move? method pass all tests Greg wrote.